split categories component from sidebar
